Class 'examinee': a single examinee

examinee is an S4 class to represent a single examinee.

Slots

examinee_id

the ID of examinee.

current_grade

the current grade the examinee is in.

current_phase

the current phase the examinee is in.

current_test

the current test the examinee is in.

current_module

the current module the examinee is in.

grade_log

grades that the examinee belonged at each module position.

phase_log

phases that the examinee belonged at each module position.

test_log

tests that the examinee belonged at each module position.

module_log

modules that the examinee belonged at each module position.

n_module

the number of modules the examinee received. This is the number of module positions.

true_theta

a vector containing the true theta (if simulated) of the examinee, for each module position.

initial_theta_in_module

a vector containing initial thetas used in each module.

prior_par_by_module

a list containing prior parameters used for each module.

estimated_theta_by_phase

a list containing estimated thetas and SEs using items in each phase.

estimated_theta_by_test

a list containing estimated thetas and SEs using combined items in each test.

estimated_theta_for_routing

a list containing estimated thetas and SEs that were used for routing.

alpha

the alpha value used to compute lower and upper bounds.

selection_theta

a list containing selection thetas in each module position.

interim_theta

a list containing interim thetas and SEs in each module position.

administered_items

a list containing administered items in each module position.

administered_stimuli

a list containing administered stimuli in each module position.

response

a list containing the examinee response in each module position.

item_data

a list containing item_pool of administered items.

routing_based_on

a vector containing the routing was based on estimated_theta_by_phase or estimated_theta_by_test at each module position.