Update a constraints object to exclude administered items

The function excludeAdministeredItems produces a new constraints object that excludes administered items from being selected.

excludeAdministeredItems(constraints, administered_items)

Arguments

constraints
a constraints object.

administered_items
item names of previously administered items.

Value

a constraints object that also constrains the administered items to be excluded.

Examples

if (FALSE) {
require(TestDesign)

cfg <- createShadowTestConfig(
  MIP = list(solver = "lpsymphony")
)
constraints <- constraints_reading
solution <- Shadow(cfg, constraints, true_theta = 0)
administered_items <- solution@output[[1]]@administered_item_index
administered_items <- solution@constraints@pool@id[administered_items]
administered_items

updated_constraints <- excludeAdministeredItems(constraints, administered_items)

solution <- Shadow(cfg, updated_constraints, true_theta = 0)
administered_items <- solution@output[[1]]@administered_item_index
administered_items <- solution@constraints@pool@id[administered_items]
administered_items ## entirely different from above
}