Extension of plot()

Extension of plot()

# S4 method for output_maat
plot(
  x,
  y,
  type,
  examinee_id = 1,
  cut_scores = NULL,
  theta_range = c(-4, 4),
  main = NULL,
  box_color = "PaleTurquoise"
)

Arguments

x
x

y
y

type
the type of plot. route plots the number of examinees routed to each path across the course of entire assessment. correlation produces a scatterplot of thetas across administrations. audit plots interim thetas over modules for a single examinee.

examinee_id
the examinee ID to plot.

cut_scores
(optional) a named list containing cut scores for each grade.

theta_range
the theta range to use in scatter plots when x is an examinee list.

main
the figure title to use in scatter plots when x is an examinee list.

box_color
the cell color to use when type is route. (default = PaleTurquoise)

Value

the route plot.

Examples

# \donttest{
library(TestDesign)
config <- createShadowTestConfig(
  final_theta = list(
    method = "MLE"
  )
)
examinee_list <- maat(
  examinee_list          = examinee_list_math,
  assessment_structure   = assessment_structure_math,
  module_list            = module_list_math,
  overlap_control_policy = "all",
  transition_CI_alpha    = 0.05,
  config                 = config,
  cut_scores             = cut_scores_math
)

plot(examinee_list, type = "route")
plot(examinee_list, type = "correlation")
plot(examinee_list, type = "audit", examinee_id = 1)

# }