Plot module routes

plotModuleRoutes is a function for building a chart for module routes.

plotModuleRoutes(
  examinee_list,
  assessment_structure,
  examinee_id = "all",
  font_size = 15,
  box_colour = "PaleTurquoise"
)

Arguments

examinee_list
the output from maat.

assessment_structure
an assessment_structure object.

examinee_id
the examinee ID to plot the module route. all plots the count that each routing node was used. (default = all)

font_size
the font size for the count labels. (default = 15)

box_colour
the colour for the boxes. (default = PaleTurquoise)