Update the grade slot of an examinee object

updateGrade is a function for determining the grade an examinee is routed to.

updateGrade(
  examinee_object,
  assessment_structure,
  module_position,
  cut_scores,
  transition_policy = "CI",
  transition_CI_alpha,
  transition_percentile_lower,
  transition_percentile_upper,
  item_pool
)

Arguments

examinee_object

an examinee object.

assessment_structure

an assessment_structure object.

module_position

the current module position, ranging from 1 to 6.

cut_scores

a named list containing cut scores to be used in each grade. Each element must be named in the form G?, where ? is a number.

transition_policy

CI uses the confidence interval to perform routing.

pool_difficulty_percentile uses item difficulty percentiles of all items in the item_pool argument to perform routing.

pool_difficulty_percentile_exclude_administered uses item difficulty percentiles of all items in the item_pool argument to perform routing, excluding all previous items administered to the examinee.

on_grade does not permit any transition.

(default = CI)

transition_CI_alpha

the alpha level used when transition_policy == "CI".

transition_percentile_lower

the percentile value (between 0 and 1) used for the lower routing in percentile-based transition policies.

transition_percentile_upper

the percentile value (between 0 and 1) used for the upper routing in percentile-based transition policies.

item_pool

the item_pool object to determine difficulty range in percentile-based transition policies.

Value

an examinee object with its current_grade slot updated.

Details

Currently the routing rules are hard-coded in the function. See the vignette for a description of routing rules.