Simulate multi-stage adaptive test

maat is the main function for simulating a multi-stage adaptive test.

maat(
  examinee_list = examinee_list,
  assessment_structure,
  module_list,
  config,
  cut_scores,
  overlap_control_policy,
  transition_policy = "CI",
  combine_policy = "conditional",
  transition_CI_alpha = NULL,
  transition_percentile_lower = NULL,
  transition_percentile_upper = NULL,
  initial_theta_list = NULL,
  verbose = TRUE
)

Arguments

examinee_list
an examinee list from simExaminees.

assessment_structure
a assessment_structure object.

module_list
a module list from loadModules.

config
a config_Shadow object.

cut_scores
a named list containing cut scores to be used in each grade. Each element must be named in the form G?, where ? is a number.

overlap_control_policy
overlap control is performed by excluding administered items from being administered again within the same examinee.

all performs overlap control at all module positions.

within_test performs overlap control only within each test.

none does not perform overlap control.

transition_policy

CI uses the confidence interval to perform routing.

pool_difficulty_percentile uses item difficulty percentiles of all items in the item_pool argument to perform routing.

pool_difficulty_percentile_exclude_administered uses item difficulty percentiles of all items in the item_pool argument to perform routing, excluding all previous items administered to the examinee.

on_grade does not permit any transition.

(default = CI)

combine_policy

This is only applied when module_position %% 2 == 0 (at Phase 2, which is the end of each test).

conditional uses the combined theta (using items from the previous module combined with the current module), if the examinee was in the same grade in Phases 1 and 2. If the examinee was in different grades in Phases 1 and 2, then the theta estimate from Phase 2 is used.

always uses the combined theta.

never uses the theta estimate from Phase 2.

(default = conditional)

transition_CI_alpha
the alpha level to use when transition_policy == "CI".

transition_percentile_lower
the percentile value (between 0 and 1) to use for the lower routing when transition_policy == "difficulty_percentile".

transition_percentile_upper
the percentile value (between 0 and 1) to use for the upper routing when transition_policy == "difficulty_percentile".

initial_theta_list
(optional) a list containing initial thetas to use in each module position.

verbose
if TRUE, print status messages. (default = TRUE)

Value

a named list containing examinee objects from the simulation.

Examples

library(TestDesign)
config <- createShadowTestConfig(
  final_theta = list(
    method = "MLE"
  )
)
examinee_list <- maat(
  examinee_list          = examinee_list_math,
  assessment_structure   = assessment_structure_math,
  module_list            = module_list_math,
  overlap_control_policy = "all",
  transition_CI_alpha    = 0.05,
  config                 = config,
  cut_scores             = cut_scores_math
)